requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
neo4j==5.20.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.10.6
python-multipart==0.0.9